+++ convert/strings.go
@@ -6,7 +6,6 @@ import (
 	"encoding/hex"
 	"math"
 	"strings"
-	"unsafe"
 )
 
 // MD5Hash for md5 hash string
@@ -14,25 +13,6 @@ func MD5Hash(text string) string {
 	hasher := md5.New()
 	hasher.Write([]byte(text))
 	return hex.EncodeToString(hasher.Sum(nil))
-}
-
-// BytesToStr converts byte slice to a string without memory allocation.
-// See https://groups.google.com/forum/#!msg/Golang-Nuts/ENgbUzYvCuU/90yGx7GUAgAJ .
-//
-// Note it may break if string and/or slice header will change
-// in the future go versions.
-func BytesToStr(b []byte) string {
-	return *(*string)(unsafe.Pointer(&b))
-}
-
-// StrToBytes converts string to byte slice without a memory allocation.
-func StrToBytes(s string) (b []byte) {
-	return *(*[]byte)(unsafe.Pointer(
-		&struct {
-			string
-			Cap int
-		}{s, len(s)},
-	))
 }
 
 // SnakeCasedName convert String into Snake Case
