+++ convert/strings_1.19.go
@@ -1,0 +1,23 @@
+//go:build !go1.20
+// +build !go1.20
+
+package convert
+
+// BytesToStr converts byte slice to a string without memory allocation.
+// See https://groups.google.com/forum/#!msg/Golang-Nuts/ENgbUzYvCuU/90yGx7GUAgAJ .
+//
+// Note it may break if string and/or slice header will change
+// in the future go versions.
+func BytesToStr(b []byte) string {
+	return *(*string)(unsafe.Pointer(&b))
+}
+
+// StrToBytes converts string to byte slice without a memory allocation.
+func StrToBytes(s string) (b []byte) {
+	return *(*[]byte)(unsafe.Pointer(
+		&struct {
+			string
+			Cap int
+		}{s, len(s)},
+	))
+}
